LayoutTests/storage/indexeddb/modern/get-keyrange-expected.txt

This test exercises IDBObjectStore.get() with an IDBKeyRange as the parameter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Success getting keyRange [-1 (Closed), undefined (Open)]
ALERT: Result is "0"
ALERT: Success getting keyRange [-1 (Open), undefined (Open)]
ALERT: Result is "0"
ALERT: Success getting keyRange [0 (Closed), undefined (Open)]
ALERT: Result is "0"
ALERT: Success getting keyRange [0 (Open), undefined (Open)]
ALERT: Result is "1"
ALERT: Success getting keyRange [0.1 (Closed), undefined (Open)]
ALERT: Result is "1"
ALERT: Success getting keyRange [0.1 (Open), undefined (Open)]
ALERT: Result is "1"
ALERT: Success getting keyRange [99 (Closed), undefined (Open)]
ALERT: Result is "99"
ALERT: Success getting keyRange [99 (Open), undefined (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [99.1 (Closed), undefined (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [99.1 (Open), undefined (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [undefined (Open), 100 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 100 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 99 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 99 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 98.99999 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 98.99999 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 98 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 98 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 0 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), 0 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), -0.1 (Closed)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [undefined (Open), -0.1 (Open)]
ALERT: Result is NegInf
ALERT: Success getting keyRange [2.5 (Closed), 3.5 (Closed)]
ALERT: Result is "3"
ALERT: Success getting keyRange [-0.5 (Closed), 0.5 (Closed)]
ALERT: Result is "0"
ALERT: Success getting keyRange [98.5 (Closed), 99.5 (Closed)]
ALERT: Result is "99"
ALERT: Success getting keyRange [-1 (Closed), 0 (Closed)]
ALERT: Result is "0"
ALERT: Success getting keyRange [-1 (Open), 0 (Closed)]
ALERT: Result is "0"
ALERT: Success getting keyRange [-1 (Closed), 0 (Open)]
ALERT: Result is undefined
ALERT: Success getting keyRange [-1 (Open), 0 (Open)]
ALERT: Result is undefined
ALERT: Success getting keyRange [3 (Closed), 4 (Closed)]
ALERT: Result is "3"
ALERT: Success getting keyRange [3 (Open), 4 (Closed)]
ALERT: Result is "4"
ALERT: Success getting keyRange [3 (Closed), 4 (Open)]
ALERT: Result is "3"
ALERT: Success getting keyRange [3 (Open), 4 (Open)]
ALERT: Result is undefined
ALERT: Success getting keyRange [99 (Closed), 100 (Closed)]
ALERT: Result is "99"
ALERT: Success getting keyRange [99 (Open), 100 (Closed)]
ALERT: Result is undefined
ALERT: Success getting keyRange [99 (Closed), 100 (Open)]
ALERT: Result is "99"
ALERT: Success getting keyRange [99 (Open), 100 (Open)]
ALERT: Result is undefined
ALERT: Success getting keyRange [Infinity (Closed), a (Closed)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Closed)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed), a (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Open)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [AS (Closed), a (Closed)]
ALERT: Result is AS
ALERT: Success getting keyRange [AS (Open), a (Closed)]
ALERT: Result is As
ALERT: Success getting keyRange [AS (Closed), a (Open)]
ALERT: Result is AS
ALERT: Success getting keyRange [AS (Open), a (Open)]
ALERT: Result is As
ALERT: Success getting keyRange [Infinity (Closed),  (Closed)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open),  (Closed)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed),  (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open),  (Open)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed), a (Closed)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Closed)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed), a (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Open)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed), a (Closed)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Closed)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Infinity (Closed), a (Open)]
ALERT: Result is PosInf
ALERT: Success getting keyRange [Infinity (Open), a (Open)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Fri Nov 04 1955 17:00:00 GMT-0700 (PDT) (Closed), Wed Oct 21 2015 09:00:00 GMT-0700 (PDT) (Closed)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Fri Nov 04 1955 17:00:00 GMT-0700 (PDT) (Open), Wed Oct 21 2015 09:00:00 GMT-0700 (PDT) (Closed)]
ALERT: Result is Fish under the sea
ALERT: Success getting keyRange [Fri Nov 04 1955 17:00:00 GMT-0700 (PDT) (Closed), Wed Oct 21 2015 09:00:00 GMT-0700 (PDT) (Open)]
ALERT: Result is Flux capacitor
ALERT: Success getting keyRange [Fri Nov 04 1955 17:00:00 GMT-0700 (PDT) (Open), Wed Oct 21 2015 09:00:00 GMT-0700 (PDT) (Open)]
ALERT: Result is Fish under the sea
ALERT: readonly transaction complete
PASS successfullyParsed is true

TEST COMPLETE
